# src/routes/letra.js
const express = require('express');
const router = express.Router();
const authMiddleware = require('../middlewares/authMiddleware');
const letraSchema = require('../models/letra');

const { updateCantidadLetras, updateValorNominalTotal } = require('../services/cartera');

router.post('/create-letra', authMiddleware, async (req, res) => {
    try {
        const letra = new letraSchema({
            ...req.body,
            fecha_emision: req.body.fecha_emision, // El formato debe ser YYYY-MM-DD
            fecha_descuento: req.body.fecha_descuento, // El formato debe ser YYYY-MM-DD
            fecha_vencimiento: req.body.fecha_vencimiento // El formato debe ser YYYY-MM-DD
        });

        const newLetra = await letra.save();
        await updateCantidadLetras(letra.carteraId, 1);
        await updateValorNominalTotal(letra.carteraId, letra.valor_nominal);
        res.json(newLetra);
    } catch (error) {
        res.json(error.message);
    }
});

router.get('/letras/:id', authMiddleware, async (req, res) => {
    try {
        const letras = await letraSchema.find({ carteraId: req.params.id });
        res.json(letras);
    } catch (error) {
        res.json(error.message);
    }
});

router.get('/letra/:id', authMiddleware, async (req, res) => {
    try {
        const letra = await letraSchema.find({ _id: req.params.id });
        res.json(letra);
    } catch (error) {
        res.json(error.message);
    }
});

router.delete('/letra/:id', authMiddleware, async (req, res) => {
    try {
        const letra = await letraSchema.findByIdAndDelete(req.params.id);
        if (!letra) {
            return res.status(404).json({ message: 'Letra not found' });
        }
        await updateCantidadLetras(letra.carteraId, -1);
        await updateValorNominalTotal(letra.carteraId, -letra.valor_nominal);
        res.status(200).json({ message: 'Letra deleted successfully' });
    } catch (error) {
        res.status(500).json(error.message);
    }
});

module.exports = router;